+++ spec/test_prof/cops/rspec/aggregate_failures_spec.rb
@@ -29,6 +29,16 @@ describe RuboCop::Cop::RSpec::AggregateFailures, :config do
     expect(cop.messages.first).to eq('Use :aggregate_failures instead of several one-liners.')
   end
 
+  it 'rejects two its one-liners in a row' do
+    inspect_source(['context "request" do',
+                    '  its(:status) { is_expected.to eq 200 }',
+                    '  its(:body) { is_expected.to eq "OK" }',
+                    'end'])
+
+    expect(cop.offenses.size).to eq(1)
+    expect(cop.messages.first).to eq('Use :aggregate_failures instead of several one-liners.')
+  end
+
   it 'rejects two one-liners when blank lines and non-example blocks' do
     inspect_source(['context "request" do',
                     '  let(:user) { create(:user) } ',
@@ -43,7 +53,7 @@ describe RuboCop::Cop::RSpec::AggregateFailures, :config do
     expect(cop.messages.first).to eq('Use :aggregate_failures instead of several one-liners.')
   end
 
-  it 'rejects one-liners with nested context' do
+  it 'accepts one-liners with nested context' do
     inspect_source(['context "request" do',
                     '  it { is_expected.to be_success }',
                     '  it "works" do',
@@ -66,9 +76,16 @@ describe RuboCop::Cop::RSpec::AggregateFailures, :config do
     expect(cop.offenses).to be_empty
   end
 
+  it 'accepts single its one-liner' do
+    inspect_source(['context "request" do',
+                    '  its(:status) { is_expected.to eq 200 }',
+                    'end'])
+    expect(cop.offenses).to be_empty
+  end
+
   it 'accepts non-regular one-liners' do
     inspect_source(['context "request" do',
-                    '  its(:foo) { is_expected.to be_success }',
+                    '  xit { is_expected.to be_success }',
                     '  pending { is_expected.to fail }',
                     'end'])
     expect(cop.offenses).to be_empty
@@ -103,6 +120,17 @@ describe RuboCop::Cop::RSpec::AggregateFailures, :config do
     expect(cop.offenses).to be_empty
   end
 
+  it 'handles its edge cases' do
+    inspect_source(['context "request" do',
+                    '  its(:status) { is_expected.to eq 200 }',
+                    '  its("body.size") { is_expected.to eq 100 }',
+                    '  its([:size, "count"]) { is_expected.to eq [100, 100] }',
+                    '  its(:body) { is_expected.to be_json_eql(%({"data":{"status":"OK}})).excluding("id") }',
+                    'end'])
+    expect(cop.offenses.size).to eq(1)
+    expect(cop.messages.first).to eq('Use :aggregate_failures instead of several one-liners.')
+  end
+
   describe "#autocorrect" do
     it "corrects two one-liners" do
       new_source = autocorrect_source(
@@ -122,6 +150,24 @@ describe RuboCop::Cop::RSpec::AggregateFailures, :config do
       )
     end
 
+    it "corrects two its one-liners" do
+      new_source = autocorrect_source(
+        ['context "request" do',
+         '  its(:status) { is_expected.to eq 200 }',
+         '  its(:body) { is_expected.to eq "OK" }',
+         'end']
+      )
+
+      expect(new_source).to eq(
+        ['context "request" do',
+         '  it "works", :aggregate_failures do',
+         '    expect(subject.status).to eq 200',
+         '    expect(subject.body).to eq "OK"',
+         '  end',
+         'end'].join("\n")
+      )
+    end
+
     it 'corrects indented one-liners when blank lines and non-example blocks' do
       new_source = autocorrect_source(
         ['describe "GET #index" do',
@@ -133,6 +179,9 @@ describe RuboCop::Cop::RSpec::AggregateFailures, :config do
          '      ',
          '      ',
          '    it { expect(response.body).to eq "OK" }',
+         '      ',
+         '      ',
+         '    its(:status) { is_expected.to eq 200 }',
          '  end',
          'end']
       )
@@ -145,6 +194,7 @@ describe RuboCop::Cop::RSpec::AggregateFailures, :config do
          '    it "works", :aggregate_failures do',
          '      is_expected.to be_success',
          '      expect(response.body).to eq "OK"',
+         '      expect(subject.status).to eq 200',
          '    end',
          '  end',
          'end'].join("\n")
@@ -163,10 +213,14 @@ describe RuboCop::Cop::RSpec::AggregateFailures, :config do
           '      ',
           '      ',
           '    it { expect(response.body).to eq "OK" }',
+          '      ',
+          '      ',
+          '    its(:status) { is_expected.to eq 200 }',
           '',
           '    context "sub-request", :invalid do',
           '      it { is_expected.not_to be_success }',
           '      it { expect(response.body).to eq "FAILED" }',
+          '      its(:status) { is_expected.to eq 404 }',
           '    end',
           '  end',
           'end'
@@ -182,16 +236,44 @@ describe RuboCop::Cop::RSpec::AggregateFailures, :config do
          '    it "works", :aggregate_failures do',
          '      is_expected.to be_success',
          '      expect(response.body).to eq "OK"',
+         '      expect(subject.status).to eq 200',
          '    end',
          '',
          '    context "sub-request", :invalid do',
          '      it "works", :aggregate_failures do',
          '        is_expected.not_to be_success',
          '        expect(response.body).to eq "FAILED"',
+         '        expect(subject.status).to eq 404',
          '      end',
          '    end',
          '  end',
          'end'].join("\n")
+      )
+    end
+
+    it "corrects its edge cases" do
+      new_source = autocorrect_source(
+        [
+          'context "request" do',
+          '  its(:status) { is_expected.to eq 200 }',
+          '  its("body.size") { is_expected.to eq 100 }',
+          '  its([:size, "count"]) { is_expected.to eq [100, 100] }',
+          '  its(:body) { is_expected.to be_json_eql(%({"data":{"status":"OK}})).excluding("id") }',
+          'end'
+        ]
+      )
+
+      expect(new_source).to eq(
+        [
+          'context "request" do',
+          '  it "works", :aggregate_failures do',
+          '    expect(subject.status).to eq 200',
+          '    expect(subject.body.size).to eq 100',
+          '    expect(subject.dig(size, count)).to eq [100, 100]',
+          '    expect(subject.body).to be_json_eql(%({"data":{"status":"OK}})).excluding("id")',
+          '  end',
+          'end'
+        ].join("\n")
       )
     end
   end